Updated lib, command line application, and cmakelist files

+++ controller/lib/src/audio_unit_descriptor_imp.cpp
@@ -38,7 +38,7 @@
 
 namespace avdecc_lib
 {
-    audio_unit_descriptor_imp::audio_unit_descriptor_imp(end_station_imp *end_station_obj, const uint8_t *frame, size_t pos, size_t frame_len) : descriptor_base_imp(end_station_obj)
+    audio_unit_descriptor_imp::audio_unit_descriptor_imp(end_station_imp *end_station_obj, const uint8_t *frame, ssize_t pos, size_t frame_len) : descriptor_base_imp(end_station_obj)
     {
         desc_audio_read_returned = jdksavdecc_descriptor_audio_read(&audio_unit_desc, frame, pos, frame_len);
 
@@ -260,7 +260,7 @@ namespace avdecc_lib
         return audio_unit_desc.current_sampling_rate;
     }
 
-    uint32_t STDCALL audio_unit_descriptor_imp::get_sampling_rate_by_index(uint32_t sampling_rate_index)
+    uint32_t STDCALL audio_unit_descriptor_imp::get_sampling_rate_by_index(size_t sampling_rate_index)
     {
         return sample_rates_vec.at(sampling_rate_index);
     }
@@ -290,7 +290,7 @@ namespace avdecc_lib
     {
         struct jdksavdecc_frame *cmd_frame;
         struct jdksavdecc_aem_command_set_sampling_rate aem_cmd_set_sampling_rate;
-        int aem_cmd_set_sampling_rate_returned;
+        ssize_t aem_cmd_set_sampling_rate_returned;
         cmd_frame = (struct jdksavdecc_frame *)malloc(sizeof(struct jdksavdecc_frame));
 
         /******************************************* AECP Common Data **********************************************/
@@ -325,10 +325,10 @@ namespace avdecc_lib
 
     }
 
-    int audio_unit_descriptor_imp::proc_set_sampling_rate_resp(void *&notification_id, const uint8_t *frame, uint16_t frame_len, int &status)
+    int audio_unit_descriptor_imp::proc_set_sampling_rate_resp(void *&notification_id, const uint8_t *frame, size_t frame_len, int &status)
     {
         struct jdksavdecc_frame *cmd_frame;
-        int aem_cmd_set_sampling_rate_resp_returned;
+        ssize_t aem_cmd_set_sampling_rate_resp_returned;
         uint32_t msg_type;
         bool u_field;
 
@@ -361,7 +361,7 @@ namespace avdecc_lib
     {
         struct jdksavdecc_frame *cmd_frame;
         struct jdksavdecc_aem_command_get_sampling_rate aem_cmd_get_sampling_rate;
-        int aem_cmd_get_sampling_rate_returned;
+        ssize_t aem_cmd_get_sampling_rate_returned;
         cmd_frame = (struct jdksavdecc_frame *)malloc(sizeof(struct jdksavdecc_frame));
 
         /******************************************** AECP Common Data *********************************************/
@@ -395,10 +395,10 @@ namespace avdecc_lib
     }
 
 
-    int audio_unit_descriptor_imp::proc_get_sampling_rate_resp(void *&notification_id, const uint8_t *frame, uint16_t frame_len, int &status)
+    int audio_unit_descriptor_imp::proc_get_sampling_rate_resp(void *&notification_id, const uint8_t *frame, size_t frame_len, int &status)
     {
         struct jdksavdecc_frame *cmd_frame;
-        int aem_cmd_get_sampling_rate_resp_returned;
+        ssize_t aem_cmd_get_sampling_rate_resp_returned;
         uint32_t msg_type;
         bool u_field;
