+++ controller/lib/src/stream_input_descriptor_imp.cpp
@@ -40,7 +40,7 @@
 
 namespace avdecc_lib
 {
-    stream_input_descriptor_imp::stream_input_descriptor_imp(end_station_imp *end_station_obj, const uint8_t *frame, size_t pos, size_t frame_len) : descriptor_base_imp(end_station_obj)
+    stream_input_descriptor_imp::stream_input_descriptor_imp(end_station_imp *end_station_obj, const uint8_t *frame, ssize_t pos, size_t frame_len) : descriptor_base_imp(end_station_obj)
     {
         stream_input_desc_read_returned = jdksavdecc_descriptor_stream_read(&stream_input_desc, frame, pos, frame_len);
 
@@ -321,7 +321,7 @@ namespace avdecc_lib
     {
         struct jdksavdecc_frame *cmd_frame;
         struct jdksavdecc_aem_command_set_stream_format aem_cmd_set_stream_format;
-        int aem_cmd_set_stream_format_returned;
+        ssize_t aem_cmd_set_stream_format_returned;
         cmd_frame = (struct jdksavdecc_frame *)malloc(sizeof(struct jdksavdecc_frame));
 
         /******************************************** AECP Common Data *********************************************/
@@ -355,10 +355,10 @@ namespace avdecc_lib
         return 0;
     }
 
-    int stream_input_descriptor_imp::proc_set_stream_format_resp(void *&notification_id, const uint8_t *frame, uint16_t frame_len, int &status)
+    int stream_input_descriptor_imp::proc_set_stream_format_resp(void *&notification_id, const uint8_t *frame, size_t frame_len, int &status)
     {
         struct jdksavdecc_frame *cmd_frame;
-        int aem_cmd_set_stream_format_resp_returned;
+        ssize_t aem_cmd_set_stream_format_resp_returned;
         uint32_t msg_type;
         bool u_field;
 
@@ -391,7 +391,7 @@ namespace avdecc_lib
     {
         struct jdksavdecc_frame *cmd_frame;
         struct jdksavdecc_aem_command_get_stream_format aem_cmd_get_stream_format;
-        int aem_cmd_get_stream_format_returned;
+        ssize_t aem_cmd_get_stream_format_returned;
         cmd_frame = (struct jdksavdecc_frame *)malloc(sizeof(struct jdksavdecc_frame));
 
         /******************************************** AECP Common Data *********************************************/
@@ -424,10 +424,10 @@ namespace avdecc_lib
         return 0;
     }
 
-    int stream_input_descriptor_imp::proc_get_stream_format_resp(void *&notification_id, const uint8_t *frame, uint16_t frame_len, int &status)
+    int stream_input_descriptor_imp::proc_get_stream_format_resp(void *&notification_id, const uint8_t *frame, size_t frame_len, int &status)
     {
         struct jdksavdecc_frame *cmd_frame;
-        int aem_cmd_get_stream_format_resp_returned;
+        ssize_t aem_cmd_get_stream_format_resp_returned;
         uint32_t msg_type;
         bool u_field;
 
@@ -463,7 +463,7 @@ namespace avdecc_lib
         return 0;
     }
 
-    int stream_input_descriptor_imp::proc_set_stream_info_resp(void *&notification_id, const uint8_t *frame, uint16_t frame_len, int &status)
+    int stream_input_descriptor_imp::proc_set_stream_info_resp(void *&notification_id, const uint8_t *frame, size_t frame_len, int &status)
     {
         log_imp_ref->post_log_msg(LOGGING_LEVEL_ERROR, "Need to implement SET_STREAM_INFO response.");
 
@@ -474,7 +474,7 @@ namespace avdecc_lib
     {
         struct jdksavdecc_frame *cmd_frame;
         struct jdksavdecc_aem_command_get_stream_info aem_cmd_get_stream_info;
-        int aem_cmd_get_stream_info_returned;
+        ssize_t aem_cmd_get_stream_info_returned;
         cmd_frame = (struct jdksavdecc_frame *)malloc(sizeof(struct jdksavdecc_frame));
 
         /******************************************** AECP Common Data *******************************************/
@@ -507,10 +507,10 @@ namespace avdecc_lib
         return 0;
     }
 
-    int stream_input_descriptor_imp::proc_get_stream_info_resp(void *&notification_id, const uint8_t *frame, uint16_t frame_len, int &status)
+    int stream_input_descriptor_imp::proc_get_stream_info_resp(void *&notification_id, const uint8_t *frame, size_t frame_len, int &status)
     {
         struct jdksavdecc_frame *cmd_frame;
-        int aem_cmd_get_stream_info_resp_returned;
+        ssize_t aem_cmd_get_stream_info_resp_returned;
         uint32_t msg_type;
         bool u_field;
 
@@ -542,7 +542,7 @@ namespace avdecc_lib
     {
         struct jdksavdecc_frame *cmd_frame;
         struct jdksavdecc_aem_command_start_streaming aem_cmd_start_streaming;
-        int aem_cmd_start_streaming_returned;
+        ssize_t aem_cmd_start_streaming_returned;
         cmd_frame = (struct jdksavdecc_frame *)malloc(sizeof(struct jdksavdecc_frame));
 
         /******************************************** AECP Common Data *******************************************/
@@ -575,11 +575,11 @@ namespace avdecc_lib
         return 0;
     }
 
-    int stream_input_descriptor_imp::proc_start_streaming_resp(void *&notification_id, const uint8_t *frame, uint16_t frame_len, int &status)
+    int stream_input_descriptor_imp::proc_start_streaming_resp(void *&notification_id, const uint8_t *frame, size_t frame_len, int &status)
     {
         struct jdksavdecc_frame *cmd_frame;
         struct jdksavdecc_aem_command_start_streaming_response aem_cmd_start_streaming_resp;
-        int aem_cmd_start_streaming_resp_returned;
+        ssize_t aem_cmd_start_streaming_resp_returned;
         uint32_t msg_type;
         bool u_field;
 
@@ -611,7 +611,7 @@ namespace avdecc_lib
     {
         struct jdksavdecc_frame *cmd_frame;
         struct jdksavdecc_aem_command_stop_streaming aem_cmd_stop_streaming;
-        int aem_cmd_stop_streaming_returned;
+        ssize_t aem_cmd_stop_streaming_returned;
         cmd_frame = (struct jdksavdecc_frame *)malloc(sizeof(struct jdksavdecc_frame));
 
         /******************************************* AECP Common Data *******************************************/
@@ -644,11 +644,11 @@ namespace avdecc_lib
         return 0;
     }
 
-    int stream_input_descriptor_imp::proc_stop_streaming_resp(void *&notification_id, const uint8_t *frame, uint16_t frame_len, int &status)
+    int stream_input_descriptor_imp::proc_stop_streaming_resp(void *&notification_id, const uint8_t *frame, size_t frame_len, int &status)
     {
         struct jdksavdecc_frame *cmd_frame;
         struct jdksavdecc_aem_command_stop_streaming_response aem_cmd_stop_streaming_resp;
-        int aem_cmd_stop_streaming_resp_returned;
+        ssize_t aem_cmd_stop_streaming_resp_returned;
         uint32_t msg_type;
         bool u_field;
 
@@ -680,7 +680,7 @@ namespace avdecc_lib
     {
         struct jdksavdecc_frame *cmd_frame;
         struct jdksavdecc_acmpdu acmp_cmd_connect_rx;
-        int acmp_cmd_connect_rx_returned;
+        ssize_t acmp_cmd_connect_rx_returned;
         uint64_t listener_guid = base_end_station_imp_ref->get_entity_desc_by_index(0)->entity_id();
 
         cmd_frame = (struct jdksavdecc_frame *)malloc(sizeof(struct jdksavdecc_frame));
@@ -718,10 +718,10 @@ namespace avdecc_lib
         return 0;
     }
 
-    int stream_input_descriptor_imp::proc_connect_rx_resp(void *&notification_id, const uint8_t *frame, uint16_t frame_len, int &status)
+    int stream_input_descriptor_imp::proc_connect_rx_resp(void *&notification_id, const uint8_t *frame, size_t frame_len, int &status)
     {
         struct jdksavdecc_frame *cmd_frame;
-        int acmp_cmd_connect_rx_resp_returned;
+        ssize_t acmp_cmd_connect_rx_resp_returned;
         uint32_t msg_type;
 
         cmd_frame = (struct jdksavdecc_frame *)malloc(sizeof(struct jdksavdecc_frame));
@@ -751,7 +751,7 @@ namespace avdecc_lib
     {
         struct jdksavdecc_frame *cmd_frame;
         struct jdksavdecc_acmpdu acmp_cmd_disconnect_rx;
-        int acmp_cmd_disconnect_rx_returned;
+        ssize_t acmp_cmd_disconnect_rx_returned;
         uint64_t listener_guid = base_end_station_imp_ref->get_entity_desc_by_index(0)->entity_id();
 
         cmd_frame = (struct jdksavdecc_frame *)malloc(sizeof(struct jdksavdecc_frame));
@@ -789,10 +789,10 @@ namespace avdecc_lib
         return 0;
     }
 
-    int stream_input_descriptor_imp::proc_disconnect_rx_resp(void *&notification_id, const uint8_t *frame, uint16_t frame_len, int &status)
+    int stream_input_descriptor_imp::proc_disconnect_rx_resp(void *&notification_id, const uint8_t *frame, size_t frame_len, int &status)
     {
         struct jdksavdecc_frame *cmd_frame;
-        int acmp_cmd_disconnect_rx_resp_returned;
+        ssize_t acmp_cmd_disconnect_rx_resp_returned;
         uint32_t msg_type;
 
         cmd_frame = (struct jdksavdecc_frame *)malloc(sizeof(struct jdksavdecc_frame));
@@ -822,7 +822,7 @@ namespace avdecc_lib
     {
         struct jdksavdecc_frame *cmd_frame;
         struct jdksavdecc_acmpdu acmp_cmd_get_rx_state;
-        int acmp_cmd_get_rx_state_returned;
+        ssize_t acmp_cmd_get_rx_state_returned;
         uint64_t listener_guid = base_end_station_imp_ref->get_entity_desc_by_index(0)->entity_id();
 
         cmd_frame = (struct jdksavdecc_frame *)malloc(sizeof(struct jdksavdecc_frame));
@@ -860,10 +860,10 @@ namespace avdecc_lib
         return 0;
     }
 
-    int stream_input_descriptor_imp::proc_get_rx_state_resp(void *&notification_id, const uint8_t *frame, uint16_t frame_len, int &status)
+    int stream_input_descriptor_imp::proc_get_rx_state_resp(void *&notification_id, const uint8_t *frame, size_t frame_len, int &status)
     {
         struct jdksavdecc_frame *cmd_frame;
-        int acmp_cmd_get_rx_state_resp_returned;
+        ssize_t acmp_cmd_get_rx_state_resp_returned;
         uint32_t msg_type;
 
         cmd_frame = (struct jdksavdecc_frame *)malloc(sizeof(struct jdksavdecc_frame));
